Bugfix if own subtype of afe_single_upload
Fixes a bug when creating an own form type which has afe_single_upload as a parent.

When creating a form type with afe_single_upload as parent the file on the entity is not set correctly. Instead it is set to array([file] => [UploadedFile]).
This happens because the SingleUploadSubscriber only checks the actual form field onto it's type, but not the parents. So if you extend the afe_single_upload its behavior changes because of that.

The fix just recurses over all parents and checks if the form extends from afe_single_upload.